requirements.txt
transformers
torch
torchvision
gradio
Pillow
